新增邮箱校验

+++ src/main/java/net/wendal/nutzbook/module/UserModule.java
@@ -6,21 +6,21 @@ import javax.servlet.http.HttpSession;
 
 import net.wendal.nutzbook.bean.User;
 
+import net.wendal.nutzbook.bean.UserProfile;
+import net.wendal.nutzbook.service.UserService;
+import net.wendal.nutzbook.util.Toolkit;
+import org.nutz.aop.interceptor.ioc.TransAop;
 import org.nutz.dao.Cnd;
 import org.nutz.dao.Dao;
 import org.nutz.dao.QueryResult;
 import org.nutz.dao.pager.Pager;
+import org.nutz.ioc.aop.Aop;
 import org.nutz.ioc.loader.annotation.Inject;
 import org.nutz.ioc.loader.annotation.IocBean;
 import org.nutz.lang.Strings;
 import org.nutz.lang.util.NutMap;
-import org.nutz.mvc.annotation.At;
-import org.nutz.mvc.annotation.Attr;
-import org.nutz.mvc.annotation.By;
-import org.nutz.mvc.annotation.Fail;
-import org.nutz.mvc.annotation.Filters;
-import org.nutz.mvc.annotation.Ok;
-import org.nutz.mvc.annotation.Param;
+import org.nutz.mvc.Scope;
+import org.nutz.mvc.annotation.*;
 import org.nutz.mvc.filter.CheckSession;
 
 @IocBean // 声明为Ioc容器中的一个Bean
@@ -33,12 +33,17 @@ public class UserModule {
     @Inject // 注入同名的一个ioc对象
     protected Dao dao;
 
+    @Inject protected UserService userService;
+
     @At
     public int count() { // 统计用户数的方法,算是个测试点
         return dao.count(User.class);
     }
 
-    @At
+    /**
+     *  无验证登录
+     * */
+    /*@At
     @Filters // 覆盖UserModule类的@Filter设置,因为登陆可不能要求是个已经登陆的Session
     public Object login(@Param("username")String name, @Param("password")String password, HttpSession session) {
         User user = dao.fetch(User.class, Cnd.where("name", "=", name).and("password", "=", password));
@@ -47,6 +52,35 @@ public class UserModule {
         } else {
             session.setAttribute("me", user.getId());
             return true;
+        }
+    }*/
+
+    /**
+     *  带有验证登录
+     *  请注意一下 SecurityUtils 那一行.
+     * */
+    @At
+    @Filters // 覆盖UserModule类的@Filter设置,因为登陆可不能要求是个已经登陆的Session
+    @POST
+    public Object login(@Param("username")String username,
+                        @Param("password")String password,
+                        @Param("captcha")String captcha,
+                        @Attr(scope= Scope.SESSION, value="nutz_captcha") String _captcha,
+                        HttpSession session) {
+        NutMap re = new NutMap();
+        if (!Toolkit.checkCaptcha(_captcha, captcha)) {
+            return re.setv("ok", false).setv("msg", "验证码错误");
+        }
+        
+        User user = dao.fetch(User.class, Cnd.where("name", "=", username).and("password", "=", password));
+        int userId = userService.fetch(username, password);
+        if (userId < 0) {
+            return re.setv("ok", false).setv("msg", "用户名或密码错误");
+        } else {
+            session.setAttribute("me", user.getId());
+            // 完成nutdao_realm后启用.
+            // SecurityUtils.getSubject().login(new SimpleShiroToken(userId));
+            return re.setv("ok", true);
         }
     }
 
@@ -63,30 +97,32 @@ public class UserModule {
         if (msg != null){
             return re.setv("ok", false).setv("msg", msg);
         }
-        user = dao.insert(user);
+       // user = dao.insert(user);
+        user = userService.add(user.getName(), user.getPassword());
         return re.setv("ok", true).setv("data", user);
     }
 
     @At
-    public Object update(@Param("..")User user) {
+    public Object update(@Param("password")String password, @Attr("me") int me){
         NutMap re = new NutMap();
-        String msg = checkUser(user, false);
-        if (msg != null){
-            return re.setv("ok", false).setv("msg", msg);
-        }
-        user.setName(null);// 不允许更新用户名
-        user.setCreateTime(null);//也不允许更新创建时间
-        user.setUpdateTime(new Date());// 设置正确的更新时间
-        dao.updateIgnoreNull(user);// 真正更新的其实只有password和salt
+       
+        if (Strings.isBlank(password) || password.length() < 6)
+            return new NutMap().setv("ok", false).setv("msg", "密码不符合要求");
+        userService.updatePassword(me, password);
         return re.setv("ok", true);
     }
 
+
     @At
+    @Aop(TransAop.READ_COMMITTED)
     public Object delete(@Param("id")int id, @Attr("me")int me) {
         if (me == id) {
             return new NutMap().setv("ok", false).setv("msg", "不能删除当前用户!!");
         }
+        if(id>0){
         dao.delete(User.class, id); // 再严谨一些的话,需要判断是否为>0
+        dao.clear(UserProfile.class, Cnd.where("userId", "=", me));
+        }
         return new NutMap().setv("ok", true);
     }
